libvmi: Split performance.go into cpu.go and memory.go
Split performance.go into cpu.go and memory.go to be more specific.
